chore(DTFS2-7688): reduce matrix for ukef e2e
