+++ e2e-tests/ukef/cypress/e2e/journeys/portal/submit-MIN/submit-MIN-deal-updates-portal-deal-status/submit-MIN-deal-updates-portal-deal-status.spec.js
@@ -1,0 +1,77 @@
+import relative from '../../../../relativeURL';
+import portalPages from '../../../../../../../portal/cypress/e2e/pages';
+
+import MOCK_USERS from '../../../../../../../e2e-fixtures/portal-users.fixture';
+import MOCK_MIN_DEAL_READY_TO_SUBMIT from '../../test-data/MIN-deal/dealReadyToSubmit';
+
+const { BANK1_MAKER1, BANK1_CHECKER1 } = MOCK_USERS;
+
+context('Portal to TFM deal submission', () => {
+  let deal;
+  let dealId;
+  const dealFacilities = [];
+
+  before(() => {
+    cy.insertManyDeals([MOCK_MIN_DEAL_READY_TO_SUBMIT()], BANK1_MAKER1).then((insertedDeals) => {
+      [deal] = insertedDeals;
+      dealId = deal._id;
+
+      const { mockFacilities } = deal;
+
+      cy.createFacilities(dealId, mockFacilities, BANK1_MAKER1).then((createdFacilities) => {
+        dealFacilities.push(...createdFacilities);
+      });
+    });
+  });
+
+  beforeEach(() => {
+    cy.clearCookie('dtfs-session');
+    cy.clearCookie('_csrf');
+    cy.getCookies().should('be.empty');
+  });
+
+  after(() => {
+    cy.clearCookies();
+    cy.clearCookie('dtfs-session');
+    cy.clearCookie('_csrf');
+    cy.getCookies().should('be.empty');
+  });
+
+  it('Portal MIN deal is submitted to UKEF, TFM acknowledges the submission and updates the portal deal status from `Submitted` to `Acknowledged`', () => {
+    //---------------------------------------------------------------
+    // portal maker submits deal for review
+    //---------------------------------------------------------------
+    cy.login(BANK1_MAKER1);
+    portalPages.contract.visit(deal);
+    portalPages.contract.proceedToReview().click();
+    cy.url().should('eq', relative(`/contract/${dealId}/ready-for-review`));
+
+    cy.keyboardInput(portalPages.contractReadyForReview.comments(), 'go');
+    portalPages.contractReadyForReview.readyForCheckersApproval().click();
+
+    //---------------------------------------------------------------
+    // portal checker submits deal to ukef
+    //---------------------------------------------------------------
+    cy.login(BANK1_CHECKER1);
+    portalPages.contract.visit(deal);
+
+    cy.assertText(portalPages.contract.status(), "Ready for Checker's approval");
+
+    portalPages.contract.proceedToSubmit().click();
+
+    portalPages.contractConfirmSubmission.confirmSubmit().check();
+    portalPages.contractConfirmSubmission.acceptAndSubmit().click(deal);
+
+    // expect to land on the /dashboard page with a success message
+    cy.url().should('include', '/dashboard');
+
+    //---------------------------------------------------------------
+    // portal deal status should be updated
+    //---------------------------------------------------------------
+    portalPages.contract.visit(deal);
+
+    cy.assertText(portalPages.contract.status(), 'Acknowledged');
+
+    cy.assertText(portalPages.contract.previousStatus(), 'Submitted');
+  });
+});
