+++ e2e-tests/ukef/cypress/e2e/journeys/portal/submit-deal/submit-deal-with-non-gbp-facilities/test-data/dealThatJustNeedsDates.json
