+++ e2e-tests/ukef/cypress/e2e/journeys/portal/submit-deal/submit-deal-with-unissued-facilities-and-resubmit-as-issued/submit-deal-with-unissued-loan-and-resubmit-as-issued/test-data/dealWithUnissuedLoanReadyToSubmit.js
